Network node Validation

+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/channels/processors/clients/CheckOutActorRequestProcessor.java
@@ -2,6 +2,7 @@ package com.bitdubai.fermat_p2p_plugin.layer.communications.network.node.develop
 
 import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.template.exceptions.CantDeleteRecordDataBaseException;
 import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.template.exceptions.CantInsertRecordDataBaseException;
+import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.template.exceptions.CantReadRecordDataBaseException;
 import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.template.exceptions.RecordNotFoundException;
 import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.commons.data.Package;
 import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.commons.data.client.request.CheckOutProfileMsgRequest;
@@ -151,12 +152,17 @@ public class CheckOutActorRequestProcessor extends PackageProcessor {
      * @param profileIdentity
      * @throws CantInsertRecordDataBaseException
      */
-    private void deleteCheckedInActor(String profileIdentity) throws CantDeleteRecordDataBaseException, RecordNotFoundException {
+    private void deleteCheckedInActor(String profileIdentity) throws CantDeleteRecordDataBaseException, RecordNotFoundException, CantReadRecordDataBaseException {
 
+        /*
+         * validate if exists
+         */
+        if(getDaoFactory().getCheckedInActorDao().exists(profileIdentity)) {
         /*
          * delete from the data base
          */
-        getDaoFactory().getCheckedInActorDao().delete(profileIdentity);
+            getDaoFactory().getCheckedInActorDao().delete(profileIdentity);
+        }
 
     }
